+++ packages/mixte/src/deepFn/index.ts
@@ -1,2 +1,3 @@
 export * from './src/deepFind';
+export * from './src/deepForEach';
 export * from './src/deepSome';